% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.utils.r
\name{clean.figure.folder}
\alias{clean.figure.folder}
\title{Clean figure folder}
\usage{
clean.figure.folder(figure.folder = "public/images/")
}
\arguments{
\item{fig.path}{figure folder, \strong{defaults to post.figure.path(post.filename)}}
}
\value{
nothing
}
\description{
Clean figure folder
}
\examples{
\dontrun{
clean.figure.folder()
}
}

