% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.utils.r
\name{compress.plots}
\alias{compress.plots}
\title{Compress plot(s)}
\usage{
compress.plots(post.filename, figure.folder = "public/images/",
  fig.path = post.figure.path(post.filename, figure.folder),
  pngout.location = "c:/Library/exe/pngout.exe")
}
\arguments{
\item{post.filename}{post filename}

\item{figure.folder}{figure folder, \strong{defaults to "public/images/"}}

\item{fig.path}{figure folder, \strong{defaults to post.figure.path(post.filename)}}

\item{pngout.location}{pngout location, \strong{defaults to 'c:/Library/exe/pngout.exe'}}
}
\value{
nothing
}
\description{
Compress plot(s)
}
\examples{
\dontrun{
compress.plots()
}
}

