% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.utils.r
\name{knit.post}
\alias{knit.post}
\title{Generate Markdown file given RMarkdown post file}
\usage{
knit.post(post.filename, temp.filename = post.filename,
  figure.folder = "public/images/",
  fig.path = post.figure.path(post.filename, figure.folder), base.url = "/",
  default.img.filename = "plot", fig.width = 9, fig.height = 6,
  remove.UTF = T, render.fn = render_jekyll)
}
\arguments{
\item{post.filename}{RMarkdown post filename}

\item{temp.filename}{intermidiate RMarkdown post filename, \strong{defaults to post.filename}}

\item{figure.folder}{figure folder, \strong{defaults to "public/images/"}}

\item{fig.path}{figure folder, \strong{defaults to post.figure.path(post.filename)}}

\item{base.url}{blog base url, \strong{defaults to "/"}}

\item{default.img.filename}{prefix for images generated, \strong{defaults to "plot"}}

\item{fig.width}{figure width, \strong{defaults to 9}}

\item{fig.height}{figure height, \strong{defaults to 6}}

\item{remove.UTF}{remove the non-ASCII characters, \strong{defaults to TRUE}}

\item{render.fn}{knitr render function, \strong{defaults to render_jekyll}}
}
\value{
figure folder path for given post
}
\description{
Generate Markdown file given RMarkdown post file
}
\examples{
\dontrun{
filename = 'post.test.r'
knit.post(filename)
}
}

