% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.utils.r
\name{make.rmd}
\alias{make.rmd}
\title{Convert R script to Rmd file}
\usage{
make.rmd(r.script.post.filename, out.filename = r.script.post.filename)
}
\arguments{
\item{r.script.post.filename}{R script post filename}

\item{out.filename}{output Rmd filename, \strong{defaults to r.script.post.filename}}
}
\value{
output Rmd filename
}
\description{
Convert R script to Rmd file
}
\examples{
\dontrun{
filename = 'post.test.r'
make.rmd(filename, 'post.r')
make.rmd(filename)
}
}

