% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.utils.r
\name{move.post}
\alias{move.post}
\title{Move post markdown and figures to blog folder}
\usage{
move.post(post.filename, blog.folder, figure.folder = "public/images/",
  fig.path = post.figure.path(post.filename, figure.folder),
  move.source.post = F)
}
\arguments{
\item{post.filename}{post filename}

\item{blog.folder}{blog folder}

\item{figure.folder}{figure folder, \strong{defaults to "public/images/"}}

\item{fig.path}{figure folder, \strong{defaults to post.figure.path(post.filename)}}

\item{move.source.post}{flag to indicate to move source post to blog folder, \strong{defaults to FALSE}}
}
\value{
nothing
}
\description{
Move post markdown and figures to blog folder
}
\examples{
\dontrun{
move.post()
}
}

