% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.utils.r
\name{post.figure.path}
\alias{post.figure.path}
\title{Generate figure path given post}
\usage{
post.figure.path(post.filename, figure.folder = "public/images/")
}
\arguments{
\item{post.filename}{post filename}

\item{figure.folder}{figure folder, \strong{defaults to "public/images/"}}
}
\value{
figure folder path for given post
}
\description{
Generate figure path given post
}
\examples{
\dontrun{
filename = 'post.test.r'
make.rmd(filename, 'post.r')
make.rmd(filename)
}
}

