% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.blog.r
\name{run.posts}
\alias{run.posts}
\title{Helper function used for SIT blog}
\usage{
run.posts(filename.pattern = "^\\\\d*-.*.r$", move.post = F,
  blog.folder = "web", move.source.post = F, compress.plots = F,
  pngout.location = "c:/Library/exe/pngout.exe", clean.figure.folder = T,
  add.info2post.fn = add.SIT.info2post, add.date = T, base.url = "/",
  fig.width = 9, fig.height = 6, remove.UTF = T,
  render.fn = render_jekyll)
}
\arguments{
\item{filename.pattern}{post files to process, \strong{defaults to '^\\d*-.*.r$'}}

\item{move.post}{flag to inidicate to move post to blog folder, \strong{defaults to FALSE}}

\item{blog.folder}{blog folder}

\item{move.source.post}{flag to indicate to move source post to blog folder, \strong{defaults to FALSE}}

\item{compress.plots}{flag to inidicate that plots must be compressed, \strong{defaults to FALSE}}

\item{pngout.location}{pngout location, \strong{defaults to 'c:/Library/exe/pngout.exe'}}

\item{clean.figure.folder}{flag to inidicate that figure folder must be cleaned, \strong{defaults to TRUE}}

\item{add.info2post.fn}{add custom info to the post function, \strong{defaults to add.SIT.info2post}}

\item{add.date}{flag to indicate to include date when post was run, \strong{defaults to TRUE}}

\item{base.url}{blog base url, \strong{defaults to "/"}}

\item{fig.width}{figure width, \strong{defaults to 9}}

\item{fig.height}{figure height, \strong{defaults to 6}}

\item{remove.UTF}{remove the non-ASCII characters, \strong{defaults to TRUE}}

\item{render.fn}{knitr render function, \strong{defaults to render_jekyll}}
}
\value{
output Rmd filename
}
\description{
Helper function used for SIT blog
}
\examples{
\dontrun{
filename = 'post.test.r'
make.rmd(filename, 'post.r')
make.rmd(filename)
}
}

